requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
configparser==5.2.0
pytest==7.1.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
selenium==4.4.0
webdriver-manager==3.8.3
